+++ src/main/java/com/anomalydetection/detector/heuristic/HeuristicRule.java
@@ -5,7 +5,7 @@ import com.anomalydetection.features.FeatureVector;
 @FunctionalInterface
 public interface HeuristicRule {
     RuleResult evaluate(FeatureVector vector);
-    default RuleResult evaluate(FeatureVector vector, double sensitivity) {
+    default RuleResult evaluate(FeatureVector vector, int sensitivity) {
         return evaluate(vector);
     }
     default String getRuleName() { return getClass().getSimpleName().replace("Rule", ""); }
