+++ src/main/java/com/anomalydetection/detector/heuristic/ModificationRatioRule.java
@@ -24,7 +24,7 @@ public class ModificationRatioRule implements HeuristicRule {
     }
 
     @Override
-    public RuleResult evaluate(FeatureVector vector, double sensitivity) {
+    public RuleResult evaluate(FeatureVector vector, int sensitivity) {
         double multiplier = com.anomalydetection.detector.SensitivityAdjuster.getThresholdMultiplier(sensitivity);
         double modRatio = vector.get(FeatureType.MODIFICATION_RATIO);
         double dailyOps = vector.get(FeatureType.TOTAL_OPERATIONS_NORMALIZED);
